requirements.txt
faiss-cpu
sentence-transformers
spacy
pdfplumber
rank_bm25
pdfplumber
python-docx
beautifulsoup4
pandas
openpyxl
